pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "meta-context-engineering"
version = "0.1.0"
description = "A bi-level agentic framework for evolving context engineering skills and task context"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Haoran Ye", email = "hrye@stu.pku.edu.cn"}
]

dependencies = [
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "langchain-core>=1.2.0",
    "langchain-openai>=0.2.0",
    "claude-agent-sdk>=0.1.18",
    "e2b-code-interpreter>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.5.0",
    "isort>=5.12.0",
    "ruff>=0.8.0",
]

[tool.setuptools]
packages = ["env", "mce"]

[tool.setuptools.package-data]
mce = ["py.typed"]


[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple"
default = true
